+++ components/custom_status/date_time_input.tsx
@@ -1,14 +1,15 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import React, {useEffect, useState, useCallback, CSSProperties} from 'react';
+import React, {useEffect, useState, CSSProperties} from 'react';
 import {useSelector} from 'react-redux';
-import {DayModifiers, DayPickerProps} from 'react-day-picker';
+import {DayPickerProps} from 'react-day-picker';
 import {FormattedMessage, useIntl} from 'react-intl';
 import {components} from 'react-select';
 import Creatable from 'react-select/creatable';
 import type {ActionMeta, ValueType, ControlProps, OptionsType} from 'react-select';
 import {DateTime} from 'luxon';
+import classNames from 'classnames';
 
 import moment, {Moment} from 'moment-timezone';
 
@@ -90,27 +91,29 @@ const styles = {
         ...css,
         minHeight: '40px',
         background: 'var(--center-channel-bg)',
-        border: '2px solid var(--button-bg)',
+        border: '1px solid rgba(var(--center-channel-color-rgb), 0.16)',
         boxShadow: 'none',
         '&:hover': {
-            border: '2px solid var(--button-bg)',
+            border: '1px solid rgba(var(--center-channel-color-rgb), 0.16)',
         },
-        '&:focus': {
-            border: '2px solid var(--button-bg)',
+        '&:focus-within': {
+            border: '1px solid var(--button-bg)',
+            boxShadow: 'rgba(0, 152, 255, 0.5) 0px 0px 1px 2px, rgb(0, 152, 255) 0px 0px 0px 1px',
         },
     }),
     singleValue: (css: CSSProperties) => ({
         ...css,
         color: 'var(--center-channel-color)',
     }),
+    indicatorSeparator: (style: CSSProperties) => ({
+        ...style,
+        display: 'none',
+    }),
+    menu: (style: CSSProperties) => ({
+        ...style,
+        zIndex: 1000,
+    }),
 };
-
-const placeholder = (
-    <FormattedMessage
-        id='time_dropdown.choose_time'
-        defaultMessage='Choose a time'
-    />
-);
 
 type Props = {
     time: Moment;
@@ -123,32 +126,38 @@ type Props = {
 const DateTimeInputContainer: React.FC<Props> = ({time, handleChange, timezone, onMenuChange, setIsDatePickerOpen}: Props) => {
     const locale = useSelector(getCurrentLocale);
     const [timeOptions, setTimeOptions] = useState<CreatableOption[]>([]);
+    const [selectedValue, setSelectedValue] = useState<ValueType<CreatableOption>>(null);
     const [isPopperOpen, setIsPopperOpen] = useState(false);
+    const [isMenuOpen, setIsMenuOpen] = useState(false);
     const {formatMessage} = useIntl();
 
-    const handlePopperOpenState = useCallback((isOpen: boolean) => {
+    const handlePopperOpenState = (isOpen: boolean) => {
         setIsPopperOpen(isOpen);
         setIsDatePickerOpen?.(isOpen);
-    }, []);
+    };
 
-    const handleKeyDown = useCallback((event: KeyboardEvent) => {
-        if (isKeyPressed(event, Constants.KeyCodes.ESCAPE) && isPopperOpen) {
-            handlePopperOpenState(false);
+    const handleMenuOpenState = (isOpen: boolean) => {
+        setIsMenuOpen(isOpen);
+        onMenuChange(isOpen);
+    };
+
+    const handleKeyDown = (e: React.KeyboardEvent, open: boolean, setter: (open: boolean) => void) => {
+        if (!open && isKeyPressed(e, Constants.KeyCodes.ENTER)) {
+            e.preventDefault();
+            e.stopPropagation();
+            setter(true);
         }
-    }, [isPopperOpen, handlePopperOpenState]);
-
-    useEffect(() => {
-        document.addEventListener('keydown', handleKeyDown);
-
-        return () => {
-            document.removeEventListener('keydown', handleKeyDown);
-        };
-    }, [handleKeyDown]);
+        if (open && isKeyPressed(e, Constants.KeyCodes.ESCAPE)) {
+            e.preventDefault();
+            e.stopPropagation();
+            setter(false);
+        }
+    };
 
     const setTimeAndOptions = () => {
         const currentTime = getCurrentMomentForTimezone(timezone);
         let startTime = moment(time).startOf('day');
-        if (time.date() === currentTime.date()) {
+        if (time.dayOfYear() === currentTime.dayOfYear() && time.year() === currentTime.year()) {
             startTime = getRoundedTime(currentTime);
         }
         setTimeOptions(getTimeInIntervals(startTime));
@@ -156,14 +165,19 @@ const DateTimeInputContainer: React.FC<Props> = ({time, handleChange, timezone, 
 
     useEffect(setTimeAndOptions, [time]);
 
-    const handleDayChange = (day: Date, modifiers: DayModifiers) => {
-        if (modifiers.today) {
-            const currentTime = getCurrentMomentForTimezone(timezone);
-            const roundedTime = getRoundedTime(currentTime);
+    const handleDayChange = (day: Date) => {
+        const currentMoment = getCurrentMomentForTimezone(timezone);
+        const dayWithTimezone = timezone ? moment.tz(day, timezone) : moment(day);
+        if (dayWithTimezone.isBefore(currentMoment)) {
+            const roundedTime = getRoundedTime(currentMoment);
+            const newOptions = getTimeInIntervals(roundedTime);
+            if (newOptions.length) {
+                setSelectedValue(newOptions[0]);
+            }
             handleChange(roundedTime);
         } else {
-            const dayWithTimezone = timezone ? moment.tz(day, timezone) : moment(day);
-            handleChange(dayWithTimezone.startOf('day'));
+            const newMoment = time.clone().dayOfYear(dayWithTimezone.dayOfYear()).year(dayWithTimezone.year());
+            handleChange(newMoment);
         }
         handlePopperOpenState(false);
     };
@@ -176,11 +190,12 @@ const DateTimeInputContainer: React.FC<Props> = ({time, handleChange, timezone, 
 
     const inputIcon = (
         <IconButton
-            onClick={() => handlePopperOpenState(true)}
+            onClick={() => handlePopperOpenState(!isPopperOpen)}
             icon={'calendar-outline'}
             className='dateTime__calendar-icon'
             size={'sm'}
             aria-haspopup='grid'
+            tabIndex={-1}
         />
     );
 
@@ -213,6 +228,7 @@ const DateTimeInputContainer: React.FC<Props> = ({time, handleChange, timezone, 
             newTime = setDateTime(time, value);
         }
         if (newTime.isValid()) {
+            setSelectedValue(newOption);
             handleChange(newTime);
         }
     };
@@ -237,21 +253,22 @@ const DateTimeInputContainer: React.FC<Props> = ({time, handleChange, timezone, 
         defaultMessage: 'Invalid date',
     });
 
-    const defaultTimeValue = {
-        value: time.toDate(),
-        label: (
-            <Timestamp
-                useRelative={false}
-                useDate={false}
-                value={time.toDate()}
-            />
-        ),
-    };
+    const placeholder = (
+        <Timestamp
+            useRelative={false}
+            useDate={false}
+            value={time.toDate()}
+        />
+    );
 
     return (
         <div>
-            <div className='dateTime'>
-                <div className='dateTime__date'>
+            <div className='dateTime' >
+                <div
+                    className={classNames('dateTime__date', {'dateTime__date-open': isPopperOpen})}
+                    tabIndex={0}
+                    onKeyDown={(e) => handleKeyDown(e, isPopperOpen, handlePopperOpenState)}
+                >
                     <DatePicker
                         isPopperOpen={isPopperOpen}
                         handlePopperOpenState={handlePopperOpenState}
@@ -264,7 +281,7 @@ const DateTimeInputContainer: React.FC<Props> = ({time, handleChange, timezone, 
                             readOnly={true}
                             className='dateTime__calendar-input'
                             label={localizeMessage('dnd_custom_time_picker_modal.date', 'Date')}
-                            onClick={() => handlePopperOpenState(true)}
+                            onClick={() => handlePopperOpenState(!isPopperOpen)}
                             tabIndex={-1}
                             inputPrefix={inputIcon}
                         />
@@ -275,13 +292,15 @@ const DateTimeInputContainer: React.FC<Props> = ({time, handleChange, timezone, 
                         components={{Control: CreatableControl}}
                         classNamePrefix='react-select'
                         options={timeOptions}
-                        defaultValue={defaultTimeValue}
+                        value={selectedValue}
+                        menuIsOpen={isMenuOpen}
                         onChange={handleTimeChange}
                         formatOptionLabel={formatOptionLabel}
                         isValidNewOption={isValidNewOption}
-                        onMenuOpen={() => onMenuChange(true)}
-                        onMenuClose={() => onMenuChange(false)}
+                        onMenuOpen={() => handleMenuOpenState(true)}
+                        onMenuClose={() => handleMenuOpenState(false)}
                         noOptionsMessage={noOptionsMessage}
+                        onKeyDown={(e) => handleKeyDown(e, isMenuOpen, handleMenuOpenState)}
                         placeholder={placeholder}
                         styles={styles}
                     />