+++ components/drafts/draft_actions/draft_actions.tsx
@@ -1,12 +1,17 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import React, {memo, useCallback} from 'react';
+import React, {memo} from 'react';
 import {FormattedMessage} from 'react-intl';
 import {useDispatch} from 'react-redux';
+import moment from 'moment-timezone';
 
 import {openModal} from 'actions/views/modals';
 import {ModalIdentifiers} from 'utils/constants';
+import {getCurrentMomentForTimezone} from 'utils/timezone';
+
+import SchedulePostModal from 'components/schedule_post/schedule_post_modal';
+import ScheduleIcon from 'components/widgets/icons/schedule_icon';
 
 import Action from './action';
 import SendDraftModal from './send_draft_modal';
@@ -15,21 +20,33 @@ import DeleteDraftModal from './delete_draft_modal';
 type Props = {
     displayName: string;
     draftId: string;
+    isInvalid: boolean;
+    timezone?: string;
+    isScheduled: boolean;
+    scheduleTimestamp?: number;
     onDelete: (draftId: string) => void;
     onEdit: () => void;
     onSend: (draftId: string) => void;
-}
+    onSchedule: (scheduleUTCTimestamp: number) => void;
+    onScheduleDelete: () => void;
+};
 
 function DraftActions({
     displayName,
     draftId,
+    isInvalid,
+    timezone,
+    isScheduled,
+    scheduleTimestamp,
     onDelete,
     onEdit,
     onSend,
+    onSchedule,
+    onScheduleDelete,
 }: Props) {
     const dispatch = useDispatch();
 
-    const handleDelete = useCallback(() => {
+    const handleDelete = () => {
         dispatch(openModal({
             modalId: ModalIdentifiers.DELETE_DRAFT,
             dialogType: DeleteDraftModal,
@@ -38,9 +55,9 @@ function DraftActions({
                 onConfirm: () => onDelete(draftId),
             },
         }));
-    }, [displayName]);
+    };
 
-    const handleSend = useCallback(() => {
+    const handleSend = () => {
         dispatch(openModal({
             modalId: ModalIdentifiers.SEND_DRAFT,
             dialogType: SendDraftModal,
@@ -49,22 +66,44 @@ function DraftActions({
                 onConfirm: () => onSend(draftId),
             },
         }));
-    }, [displayName]);
+    };
+
+    const handleSchedule = () => {
+        dispatch(openModal({
+            modalId: ModalIdentifiers.SCHEDULE_POST,
+            dialogType: SchedulePostModal,
+            dialogProps: {
+                timestamp: scheduleTimestamp ? moment.unix(scheduleTimestamp) : getCurrentMomentForTimezone(timezone),
+                timezone,
+                isScheduledDraft: isScheduled,
+                onConfirm: onSchedule,
+                onDelete: onScheduleDelete,
+            },
+        }));
+    };
+
+    const deleteAction = (
+        <Action
+            icon='icon-trash-can-outline'
+            id='delete'
+            name='delete'
+            tooltipText={(
+                <FormattedMessage
+                    id='drafts.actions.delete'
+                    defaultMessage='Delete draft'
+                />
+            )}
+            onClick={handleDelete}
+        />
+    );
+
+    if (isInvalid) {
+        return deleteAction;
+    }
 
     return (
         <>
-            <Action
-                icon='icon-trash-can-outline'
-                id='delete'
-                name='delete'
-                tooltipText={(
-                    <FormattedMessage
-                        id='drafts.actions.delete'
-                        defaultMessage='Delete draft'
-                    />
-                )}
-                onClick={handleDelete}
-            />
+            {deleteAction}
             <Action
                 icon='icon-pencil-outline'
                 id='edit'
@@ -76,6 +115,23 @@ function DraftActions({
                     />
                 )}
                 onClick={onEdit}
+            />
+            <Action
+                icon={ScheduleIcon}
+                id='schedule'
+                name='schedule'
+                tooltipText={scheduleTimestamp ? (
+                    <FormattedMessage
+                        id='drafts.actions.reschedule'
+                        defaultMessage='Reschedule draft'
+                    />
+                ) : (
+                    <FormattedMessage
+                        id='drafts.actions.schedule'
+                        defaultMessage='Schedule draft'
+                    />
+                )}
+                onClick={handleSchedule}
             />
             <Action
                 icon='icon-send-outline'