+++ components/reset_status_modal/reset_status_modal.tsx
@@ -12,6 +12,7 @@ import {UserStatuses} from 'utils/constants';
 import {t} from 'utils/i18n';
 import {UserStatus} from '@mattermost/types/users';
 import {PreferenceType} from '@mattermost/types/preferences';
+import './reset_status_modal.scss';
 
 t('modal.manual_status.auto_responder.message_');
 t('modal.manual_status.auto_responder.message_away');
@@ -213,6 +214,7 @@ export default class ResetStatusModal extends React.PureComponent<Props, State> 
                 title={manualStatusTitle}
                 message={manualStatusMessage}
                 confirmButtonText={manualStatusButton}
+                modalClass='resetStatus'
                 onConfirm={this.onConfirm}
                 cancelButtonText={manualStatusCancel}
                 onCancel={this.onCancel}